Completing all the loops...I think lol
I worked with Will to connect all the functions to make it work. Someone will need to run through to verify it all works properly and to touch some things up (sometimes it prints random variables -- so we could see what they do).

Also, I moved the teamList.pop to happen when the user chooses option 2 to select the away team.